requirements.txt
Django>=4.2
djangorestframework>=3.14
PyJWT>=2.8
mysqlclient>=2.1
python-dotenv>=1.0